requirements.txt
langchain
openai
python-dotenv
langchain-community
langchain-google-genai 
google-generativeai